+++ packages/core/src/utils/get-collisions.spec.ts
@@ -1,4 +1,5 @@
 import 'reflect-metadata';
+import { it, jest, describe, beforeEach, expect, xdescribe } from '@jest/globals';
 
 import { ServiceProvider } from '../types/mix';
 import { getCollisions } from './get-collisions';