+++ packages/router/src/tree.spec.ts
@@ -1,5 +1,6 @@
 import { Tree } from './tree';
 import { RouteParam } from './types';
+import { it, jest, describe, beforeEach, expect, xdescribe, beforeAll, afterEach } from '@jest/globals';
 
 class TestTree extends Tree {
   printTree(prefix = '') {